html/semantics/document-metadata/interactions-of-styling-and-scripting/script-created-link-stylesheet-does-not-block-script.html

<!DOCTYPE html>
<title>Script-created link stylesheet is not script-blocking</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
const start = performance.now();
const link = document.createElement('link');
link.rel = 'stylesheet';
link.href = 'stylesheet.py?delay=1';
document.head.appendChild(link);
</script>
<script>
// This script should not be blocked by the pending link stylesheet.
// Error margin to make the test more robust.
const epsilon = 50;
test(() => {
  const delay = performance.now() - start;
  assert_less_than(delay, 1000 - epsilon);
});
</script>
